requirements.txt
Flask==3.1.2
tensorflow==2.19.0
numpy==2.0.2
Pillow==11.0.0
gunicorn==23.0.0